pyproject.toml
[tool.poetry]
name = "final-project"
version = "0.1.0"
description = ""
authors = ["Sanjar Abrorxojaev <sabrorxojaev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.0.0"
selenium = "^4.17.2"
webdriver-manager = "^4.0.1"
allure-pytest = "^2.13.2"
faker = "^22.6.0"
pillow = "^10.2.0"
pytest-check = "^2.3.1"
gspread = "^6.0.0"
seleniumbase = "^4.23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ['.']
addopts  = ["--clean-alluredir", "--alluredir", "allure-reports"]
